Purchasing Items in the Vending Machine now works for Left Click (Buy 1 Item).

+++ src/main/java/beardlessbrady/modcurrency/block/TileEconomyBase.java
@@ -119,7 +119,7 @@ public class TileEconomyBase extends TileEntity {
         return 2;
     }
 
-    public void setLongField(int id, int value){
+    public void setLongField(int id, long value){
         switch(id){
             case FIELD_LONG_CASHRESERVE:
                 cashReserve = value;
